28th Push - Email Management
Added Email Management: This Email Management  provides a user interface for administrators to handle email communications directly from the dashboard. It allows them to send customized emails to registered users by selecting from a list of editable templates, and it includes sections for managing these templates and viewing a history of sent emails.